Move helpers to rule_helpers to avoid using BaseLinter which has things this rule doesn't need
